Changed website to weather.com 10 day display for Knoxville.
Added JS to move to element so as to show the entire 10 day forecast.

Added weather.com rader for Knoxville as second page.

Added JS to click button to hide side bar and play future radar view.

Added a Environment variable request to replace the string in app.config that points to the local user profile with a string that uses the Environment.UserName variable so it always points to the user's local profile.